fix(android): 登录/同步/统计范围 bug 修复 + 性能与资源优化
功能 Bug:
- 冷启动不再每次触发全量同步，仅首次(从未同步过)自动全量
- 登录成功后自动进入主界面，不再卡在授权 WebView
- 统计页时间范围切换真正生效(改为显式传 range)
- 同步完成后 panels 自动刷新(await 同步结束再 loadDashboard)

其他修复:
- 下拉刷新指示器空闲时不再常驻页面顶部
- 登录 WebView 离开组合时销毁，修复内存泄漏
- 聚合 token 字段 Int→Long，避免大用量溢出
- 记录页模型下拉列表缓存，翻页不再重复查询

性能/流畅度:
- 记录/会话列表改用 LazyColumn 按需组合回收
- 三张图表数据集 remember 缓存，重组不再反复重建
- loadDashboard 聚合查询并发执行，缩短首屏时间
- 自动同步定时器幂等重启，避免反复取消/重建

+++ android/app/src/main/java/io/github/yphyphyph/gogauge/data/db/UsageDao.kt
@@ -174,9 +174,9 @@ abstract class UsageDao {
             val r = byHour[h]
             HourStat(
                 hour = "%02d:00".format(h),
-                input = r?.input ?: 0,
-                output = r?.output ?: 0,
-                reasoning = r?.reasoning ?: 0,
+                input = r?.input ?: 0L,
+                output = r?.output ?: 0L,
+                reasoning = r?.reasoning ?: 0L,
             )
         }
     }
@@ -319,11 +319,11 @@ abstract class UsageDao {
                 createdAt = r.createdAt,
                 model = r.model,
                 provider = r.provider,
-                inputTokens = r.inputTokens,
-                outputTokens = r.outputTokens,
-                reasoningTokens = r.reasoningTokens,
-                cacheReadTokens = r.cacheReadTokens,
-                cacheWriteTokens = (r.cacheWrite5mTokens ?: 0) + (r.cacheWrite1hTokens ?: 0),
+                inputTokens = r.inputTokens.toLong(),
+                outputTokens = r.outputTokens.toLong(),
+                reasoningTokens = r.reasoningTokens.toLong(),
+                cacheReadTokens = r.cacheReadTokens.toLong(),
+                cacheWriteTokens = (r.cacheWrite5mTokens ?: 0).toLong() + (r.cacheWrite1hTokens ?: 0).toLong(),
                 costUsd = r.costUsd,
                 sessionId = r.sessionId,
                 plan = r.plan,
@@ -351,49 +351,49 @@ abstract class UsageDao {
     data class TotalsRow(
         @androidx.room.ColumnInfo(name = "request_count") val requestCount: Int,
         @androidx.room.ColumnInfo(name = "session_count") val sessionCount: Int,
-        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Int,
-        @androidx.room.ColumnInfo(name = "cache_hit_tokens") val cacheHitTokens: Int,
-        @androidx.room.ColumnInfo(name = "cache_write_tokens") val cacheWriteTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Int,
+        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Long,
+        @androidx.room.ColumnInfo(name = "cache_hit_tokens") val cacheHitTokens: Long,
+        @androidx.room.ColumnInfo(name = "cache_write_tokens") val cacheWriteTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Long,
         @androidx.room.ColumnInfo(name = "total_cost_usd") val totalCostUsd: Double,
     )
 
     data class DailyStatRow(
         val date: String,
-        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Int,
-        @androidx.room.ColumnInfo(name = "cache_hit_tokens") val cacheHitTokens: Int,
-        @androidx.room.ColumnInfo(name = "cache_write_tokens") val cacheWriteTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Int,
+        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Long,
+        @androidx.room.ColumnInfo(name = "cache_hit_tokens") val cacheHitTokens: Long,
+        @androidx.room.ColumnInfo(name = "cache_write_tokens") val cacheWriteTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Long,
         @androidx.room.ColumnInfo(name = "total_cost_usd") val totalCostUsd: Double,
         @androidx.room.ColumnInfo(name = "request_count") val requestCount: Int,
     )
 
-    data class HourTrendRow(val hour: Int, val input: Int, val output: Int, val reasoning: Int)
+    data class HourTrendRow(val hour: Int, val input: Long, val output: Long, val reasoning: Long)
 
     data class ModelStatRow(
         val model: String,
         @androidx.room.ColumnInfo(name = "request_count") val requestCount: Int,
         @androidx.room.ColumnInfo(name = "session_count") val sessionCount: Int,
-        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Int,
-        @androidx.room.ColumnInfo(name = "cache_hit_tokens") val cacheHitTokens: Int,
-        @androidx.room.ColumnInfo(name = "cache_write_tokens") val cacheWriteTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Int,
+        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Long,
+        @androidx.room.ColumnInfo(name = "cache_hit_tokens") val cacheHitTokens: Long,
+        @androidx.room.ColumnInfo(name = "cache_write_tokens") val cacheWriteTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Long,
         @androidx.room.ColumnInfo(name = "total_cost_usd") val totalCostUsd: Double,
     )
 
     data class SessionRow(
         @androidx.room.ColumnInfo(name = "session_id") val sessionId: String,
         @androidx.room.ColumnInfo(name = "request_count") val requestCount: Int,
-        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Int,
-        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Int,
+        @androidx.room.ColumnInfo(name = "total_input_tokens") val totalInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "uncached_input_tokens") val uncachedInputTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_output_tokens") val totalOutputTokens: Long,
+        @androidx.room.ColumnInfo(name = "total_reasoning_tokens") val totalReasoningTokens: Long,
         @androidx.room.ColumnInfo(name = "total_cost_usd") val totalCostUsd: Double,
         @androidx.room.ColumnInfo(name = "last_at") val lastAt: String,
     )
